% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap.R
\name{adj_bitmap}
\alias{adj_bitmap}
\title{Adjust bitmap picture}
\usage{
adj_bitmap(img.in, img.out, xpix.ex, ypix.ex, xpix.in, ypix.in)
}
\arguments{
\item{img.in}{Full path to the bitmap picture}

\item{img.out}{Full path to the adjusted bitmap picture}

\item{xpix.ex}{Minimum number of pixels to form a straight line in the x-axis}

\item{ypix.ex}{Minimum number of pixels to form a straight line in the y-axis}

\item{xpix.in}{Maximum number of pixels to connect lines in the x-axis}

\item{ypix.in}{Maximum number of pixels to connect lines in the y-axis}
}
\value{
A bitmap picture
}
\description{
Adjust bitmap picture
}
