% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap.R
\name{conv_bitmap}
\alias{conv_bitmap}
\title{Convert Picture to Bitmap}
\usage{
conv_bitmap(img.in, img.out)
}
\arguments{
\item{img.in}{Full path to the picture}

\item{img.out}{Full path to the converted bitmap picture}
}
\value{
A bitmap picture
}
\description{
Convert Picture to Bitmap
}
