% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{detect_rectangles}
\alias{detect_rectangles}
\title{Detect rectangles}
\usage{
detect_rectangles(img, threshold = "50\%", area.threshold = 500)
}
\arguments{
\item{img}{Full path to the image}

\item{threshold}{??}

\item{area.threshold}{Minimum size of the rectangle}
}
\value{
A dataframe
}
\description{
Detect rectangles
}
