% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap.R
\name{img_crop}
\alias{img_crop}
\title{Crop picture with Magick}
\usage{
img_crop(img.in, img.out, bound.box)
}
\arguments{
\item{img.in}{Original image}

\item{img.out}{Cropped image}

\item{bound.box}{bounding box}
}
\value{
A picture
}
\description{
Crop picture with Magick
}
