% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle_overlap.R
\name{pix_vec}
\alias{pix_vec}
\title{Generate pixel positions of rectangle in larger image}
\usage{
pix_vec(v = c(x0, y0, x1, y1, w, h), out = c("vector", "matrix"))
}
\arguments{
\item{v}{A vector containing the start- and end-positions of the rectangles
(x0, y0, x1, y1) and the width of the larger image (w)}
}
\value{
A vector or a matrix
}
\description{
Generate pixel positions of rectangle in larger image
}
\examples{
pix_vec(v = c(2,2,5,5,6,8), "vector")
pix_vec(v = c(2,2,5,5,6,8), "matrix")
}
