% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle_overlap.R
\name{rect_overlap}
\alias{rect_overlap}
\title{Calculates the overlap of rectangles in an image}
\usage{
rect_overlap(coord.list = list(c(x01, y01, x11, y11), ..., c(x0n, y0n,
  x1n, y1n)), compress = 100)
}
\arguments{
\item{coord.list}{list of coordinates for the rectangles}

\item{compress}{number of compression}
}
\value{
A dataframe
}
\description{
Calculates the overlap of rectangles in an image
}
\examples{
rect_overlap(coord.list = list(c(2,2,5,5), c(3,3,7,7), c(2,3,6,6)), compress = 0)

# check correctness
c1 <- pix_vec(c(2,2,5,5,10))
c2 <- pix_vec(c(3,3,7,7,10))
length(which(c1 \%in\% c2)) / length(c1)

}
